Finalized the C++ implementation of CI
The current version of the config.yml installs all necessary tools and builds the audio_test executable that runs all the unit
tests created for our audio engine. We are working towards getting the React implmentation finished but currently have it commented out
in the current config.yml. This is due to the massive change in project that our web application presents when compared to the
simple test-app that was created in the ar-ci-tools branch.