Re-uploaded a new extracted version of FMOD_Studio
There appears to be an issue that occurs when FMOD libraries is changed from devices that results in a linker issue.
This was appeared to be fixed in the arent007/testingCircleCI branch, but not in master. So we are adding the files
again as was done in arent007/testingCircleCI to ensure that we can continue to develop on the Raspberry Pi and have CircleCI
build properly.