Initial version of interprocess communication
This current version is using sockets to achieve interprocess communication. The server.py file should create server that allows for a UDP connection between itself and the process that will execute on s3_playback.cpp. This current version is just testing the initial connection between the two processes. Next will be inplementing the actual workflow for the server to notify the s3_playback process of which file to play.