Finished the current goal for sockets, now moving to JSON integration
The current goal for this commit was to get a working sample of socket connections that would allow for our c++ program running on the pi to request for a file to play from a python program using sockets. Currently the sample uses alot of debug messages to visualize the back and forth between the two process. The python program currently only downloads a single file and finishes execution, while the c++ program plays the a given file on repeat until the process is forced to terminate.

Things that need to be done
*Begin to work on integrating JSON downloads instead of just a file.
*Test the JSON working version before integrating the SQS queues that Brandon and Joseph have been working on.
*Implement more error checking in the python program in regards to buckets, files, and paths existing.
*See whether or not the calls for sleep and time.sleep are needed for the program to function.