Added JSON parsing and python unit tests
The current commit demonstrates a simple working JSON parser in c++ with the help of a library. This simple implementation should allow us to further implement a more robust parser for s3_playback.cpp. The current testJson.json is used to extract the file name we want to play while the server still downloads the file. The server this time around gives the s3_playback program the file name of the JSON to parse through. This implmentation should make it easier to control the various options for playback by via JSON parsing. This commit also attempts to integrate python unit tests into CircleCI. Currently the unit tests test the current functionality of the server.py file found in pyServer. By altering the config.yml we hope the current unit tests will be ran in CircleCI allowing us keep track of the server development.

*Things that still need to be done
*Tests were created for the python server, but need to imclude some unit test for the s3_playback program
*Implement more error checking on the aws buckets
*See whether the need for the time.sleep calls for the server and the sleep calls from s3_playback are necessary