Began to integrate the SQS to the python server
The current working version of the program just tests that the messages from the SQS queue are received by the python server and sent to the audio engine for parsing. Brandon is currently working on getting the parsing implemented. I will be attempting to see if the delays for each program are necessary to keep the connection between the python server and the audio engine