Added some error checking for the SQS message receiver.
Currently the addition to receivePacket.py will help check the incoming message if it contains all the necessary fields for audio playback. As well as checking that the given values meet certain conditions like the volume for playback can only consist of numeric characters, the playback option should either be true or false, and that the file name does not contain certain special characters for parsing.

*Things to do
-Comment the code receivePacket.py for clarity
-Create unit tests for the given mock in recievePacket.py(Since the actual version requires interfacing with AWS)