Added playback example that works with webapp
This simple implementation works currently with a user that I created with the web app. If you wish to run this you would need to change the listening url to a different SQS queue to receieve messages. Also hoping that this commit fixes errors that were present in the CircleCI build. Doing so will ensure that I can continue to write unit tests.

Things to do:
-Still need to write unit tests. Hopefully this commit fixes the current CircleCI issue.
-Documentation of the current python code to explain what is going on.