Added unit tests and documentation for server.py and receivePacket.py
The last commits TO DO's have been done and we have tested the communication aspects of both our server.py and receivePacket.py python code. The only thing left at this point is to integrate this implementation of the server side connection to the audio engine that will be running in s3_playback.cpp. Brandon started an implementation as Joseph is working on getting it integrated as of this commit.

TO-DO:
-See that integration of the python server and c++ audio engine is finalized and tested before moving onto working on the web applicationy