Made changes to server.py and began to add unit test for PacketParser
This commit includes some quality of life changes for the server.py such as checking if a file exists before starting a download. As well as fixing the way a volume option is passed through to allow for float values. Also altered the make file to include new unit tests for testing PacketParser.

*Things To Do
-Implement more unit test for PacketParser
-Also begin to address the issue that is listed in the Github repo dealing with inconsistant communication from the server to the audio engine