Initial working version of forked audio engine processes.
This current commit has a simple example that has two processes created in the s3_playback executable. Here we have the child process sending data over to the parent process. This simple set up should allow for us to expand upon the server connection such that the child process just deals with the server while the parent process deals with the audio engine.

*Things to do:
-Expand this example to use the server connection code.