Initial version of device registration
This initial working version allows for a JSON registration packet to be sent to TutorialTestQueue where the device waits for a queue id that represents the user's queue id to target. Currently the device has a template device.json that will be read for the queue id to use. Once a registration JSON is received, the device updates the device.json file with the queue to target and will begin to listen to the specified queue. You only need to register the device once currently and you cannot run the cloud sound executable until the device is registered

*Things to do:
-Continue to build upon the server to create pre-signed urls that the audio engine will playback.
-Think of ways to create unit tests for this device implementation and the forking implementation. (Note: Will probably have to mock the registration since they require the use of SQS)