Created the initial version of the README.md
This README.md details all the steps necessary to install and run the project in your own personal environment. However, those who download the project will have to have IAM users that are able to develop for the project. This might be an issue for people without IAM permissions who want to run the AWS CLI commands necessary.